Noticloud/Locomotive-Digital-Twin-main-2/backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.10.6
python-dotenv==1.1.1
sqlalchemy==2.0.41
psycopg2-binary==2.9.10
requests==2.32.3
PyJWT==2.12.1
prometheus-client==0.24.1
reportlab==4.4.1
websockets==16.0